requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1 ; python_version >= '3.8'
django==5.0.7
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0 ; python_version >= '3.7'
sqlparse==0.5.0 ; python_version >= '3.8'
